Refactor basketball court component and delete Team component
